Restructure code
Move all code except for main code into pkg directory.
