+++ pkg/constats.go
@@ -1,0 +1,5 @@
+package pkg
+
+const (
+	namespace = "aws_resources_exporter"
+)
